Ticket #114
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@1764 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ coupled/src/solver/cardiac/BidomainDg0Assembler.hpp
@@ -467,7 +467,7 @@ public:
         assert(fixedExtracellularPotentialNodes.size() > 0);
         for (unsigned i=0; i<fixedExtracellularPotentialNodes.size(); i++)
         {
-            if ( (int) fixedExtracellularPotentialNodes[i] >= this->mpMesh->GetNumNodes() )
+            if (fixedExtracellularPotentialNodes[i] >= this->mpMesh->GetNumNodes() )
             {
                 EXCEPTION("Fixed node number must be less than total number nodes");
             }
